modified few controllers
