+++ src/controllers/art-categories.controller.js
@@ -1,5 +1,75 @@
-const ArtCategories = require('../models/art-categories.model');
-const ArtCategoriesFormats = require('../models/art-categoriesFormats.model');
+const db = require('../utils/database');
+
+class ArtCategories {
+  constructor(category_id, name, description, margin) {
+    this.category_id = category_id;
+    this.name = name;
+    this.description = description;
+    this.margin = margin;
+  }
+
+  static fetchAll() {
+    return db.execute('SELECT * FROM category');
+  }
+
+  static post(name, description, margin) {
+    return db.execute(
+      'INSERT INTO category (name, description, margin) VALUES (?, ?, ?)',
+      [name, description, margin]
+    )
+    .then(([result]) => {
+      return { categoryId: result.insertId }; // Return the generated category ID
+    });
+  }
+
+  static update(category_id, name, description, margin) {
+    return db.execute(
+      'UPDATE category SET name = ?, description = ?, margin = ? WHERE category_id = ?',
+      [name, description, margin, category_id]
+    )
+    .then(() => {
+      return { categoryId: category_id }; // Return the updated category ID
+    });
+  }
+
+  static delete(category_id) {
+    return db.execute('DELETE FROM category WHERE category_id = ?', [category_id]);
+  }
+}
+
+class ArtCategoriesFormats {
+  constructor(format_id, category_id, format_name) {
+    this.format_id = format_id;
+    this.category_id = category_id;
+    this.format_name = format_name;
+  }
+
+  static fetchAll(category_id) {
+    return db.execute('SELECT * FROM supported_formats WHERE category_id = ?', [category_id]);
+  }
+
+  static post(category_id, format_name) {
+    return db.execute(
+      'INSERT INTO supported_formats (category_id, format_name) VALUES (?, ?)',
+      [category_id, format_name]
+    );
+  }
+
+  static update(format_id, category_id, format_name) {
+    return db.execute(
+      'UPDATE supported_formats SET category_id = ?, format_name = ? WHERE format_id = ?',
+      [category_id, format_name, format_id]
+    );
+  }
+
+  static deleteByCategoryId(category_id) {
+    return db.execute('DELETE FROM supported_formats WHERE category_id = ?', [category_id]);
+  }
+
+  static delete(format_id) {
+    return db.execute('DELETE FROM supported_formats WHERE format_id = ?', [format_id]);
+  }
+}
 
 exports.fetchAll = async (req, res, next) => {
   try {
