+++ src/controllers/artist-request.controller.js
@@ -1,8 +1,27 @@
-const User = require('../models/artist-request.model');
+
+const db = require('../utils/database');
+
+class artistRequest {
+  static getRequestedArtists() {
+    return db.execute('SELECT profile_photo_url, fName, LName, role, location, registered_at FROM user WHERE role = "artist" AND is_approved = FALSE');
+  }
+
+  static getArtistDetails(userId) {
+    return db.execute('SELECT * FROM user WHERE user_id = ?', [userId]);
+  }
+
+  static approveArtist(userId) {
+    return db.execute('UPDATE user SET is_approved = TRUE WHERE user_id = ?', [userId]);
+  }
+
+  static rejectArtist(userId) {
+    return db.execute('DELETE FROM user WHERE user_id = ?', [userId]);
+  }
+}
 
 exports.getRequestedArtists = async (req, res, next) => {
   try {
-    const requestedArtists = await User.getRequestedArtists();
+    const requestedArtists = await artistRequest.getRequestedArtists();
     res.status(200).json(requestedArtists[0]);
   } catch (error) {
     console.error('Error getting requested artists:', error);
@@ -14,7 +33,7 @@ exports.getArtistDetails = async (req, res, next) => {
   const userId = req.params.userId;
 
   try {
-    const artistDetails = await User.getArtistDetails(userId);
+    const artistDetails = await artistRequest.getArtistDetails(userId);
     res.status(200).json(artistDetails[0][0]);
   } catch (error) {
     console.error('Error getting artist details:', error);
@@ -26,7 +45,7 @@ exports.approveArtist = async (req, res, next) => {
   const userId = req.params.userId;
 
   try {
-    await User.approveArtist(userId);
+    await artistRequest.approveArtist(userId);
     res.status(200).json({ message: 'Artist approved successfully!' });
   } catch (error) {
     console.error('Error approving artist:', error);
@@ -38,11 +57,12 @@ exports.deleteAccount = async (req, res, next) => {
   const userId = req.params.userId;
 
   try {
-    await User.deleteAccount(userId);
+    await artistRequest.deleteAccount(userId);
     res.status(200).json({ message: 'Account deleted successfully!' });
   } catch (error) {
     console.error('Error deleting account:', error);
     next(error);
   }
 };
+
